reorganized spikegadgets tools into a module
